refactoring our code
